requirements.txt
dotenv==0.9.9
fastapi==0.116.1
google-genai==1.32.0
minio==7.2.16
pillow==11.3.0
pip==25.1.1
python-multipart==0.0.20
uvicorn==0.35.0
